v1.5 Editor tools

+++ maze4d/Texture.cpp
@@ -3,5 +3,5 @@
 int Texture::MIN_VALUE = 0;
 int Texture::TEX_SIZE = 16;
 int Texture::LIGHT_GRAD = 16;
-float Texture::BORDER_SIZE = 1.0f;
+int Texture::BORDER_SIZE = 1;
 bool Texture::TEX_SMOOTHERING_FLAG = false;
